+++ Airship/AirshipCore/Source/TextInput.swift
@@ -80,6 +80,7 @@ struct TextInput: View {
             createTextEditor()
                 .id(self.model.identifier)
         }
+        .padding(self.model.border?.strokeWidth ?? 0)
         #if !os(watchOS)
         .keyboardType(keyboardType)
         #endif
@@ -113,6 +114,8 @@ struct TextInput: View {
         let data = FormInputData(
             self.model.identifier,
             value: .text(trimmed.isEmpty ? nil : trimmed),
+            attributeName: self.model.attributeName,
+            attributeValue: trimmed.isEmpty ? nil : .string(trimmed),
             isValid: isValid
         )
         self.formState.updateFormInput(data)