+++ Airship/AirshipMessageCenter/Source/Views/MessageCenterListView.swift
@@ -92,14 +92,26 @@ public struct MessageCenterListView: View {
         )
 
         if #available(iOS 15.0, *) {
-            cell.listRowBackground(theme.cellColor?.adaptiveColor(for: colorScheme, darkVariation: theme.cellColorDark))
+            cell.listRowBackground(colorScheme.resolveColor(light: theme.cellColor, dark: theme.cellColorDark))
                 .listRowSeparator(
                     (theme.cellSeparatorStyle == SeparatorStyle.none)
                         ? .hidden : .automatic
                 )
-                .listRowSeparatorTint(theme.cellSeparatorColor?.adaptiveColor(for: colorScheme, darkVariation: theme.cellSeparatorColorDark))
+                .listRowSeparatorTint(colorScheme.resolveColor(light: theme.cellSeparatorColor, dark: theme.cellSeparatorColorDark))
         } else {
-            cell.listRowBackground(theme.cellColor?.adaptiveColor(for: colorScheme, darkVariation: theme.cellColorDark))
+            cell.listRowBackground(colorScheme.resolveColor(light: theme.cellColor, dark: theme.cellColorDark))
+        }
+    }
+
+    @ViewBuilder
+    private func makeCell(messageID: String) -> some View {
+        if let item = self.viewModel.messageItem(forID: messageID) {
+            makeCell(item: item, messageID: messageID)
+                .tag(messageID)
+                .opacity(self.listOpacity)
+        } else {
+            EmptyView()
+                .opacity(self.listOpacity)
         }
     }
 
@@ -107,12 +119,7 @@ public struct MessageCenterListView: View {
     private func makeList() -> some View {
         let list = List(selection: $selection) {
             ForEach(self.messageIDs, id: \.self) { messageID in
-                if let item = self.viewModel.messageItem(forID: messageID) {
-                    makeCell(item: item, messageID: messageID)
-                        .tag(messageID)
-                } else {
-                    EmptyView()
-                }
+                makeCell(messageID: messageID)
             }
             .onDelete { offsets in
                 delete(
@@ -142,6 +149,7 @@ public struct MessageCenterListView: View {
             list.refreshable {
                 await self.viewModel.refreshList()
             }
+            .disabled(self.messageIDs.isEmpty)
         } else {
             list
         }
@@ -149,17 +157,11 @@ public struct MessageCenterListView: View {
 
     @ViewBuilder
     private func makeContent() -> some View {
-        let listBackgroundColor = theme
-            .messageListBackgroundColor?
-            .adaptiveColor(
-                for: colorScheme,
-                darkVariation: theme.messageListBackgroundColorDark
-            )
+        let listBackgroundColor = colorScheme.resolveColor(light: theme.messageListBackgroundColor, dark: theme.messageListBackgroundColorDark)
 
         let content = ZStack {
             makeList()
                 .listBackground(listBackgroundColor)
-                .opacity(self.listOpacity)
                 .animation(.easeInOut(duration: 0.5), value: self.listOpacity)
                 .onChange(of: self.messageIDs) { ids in
                     if ids.isEmpty {
@@ -174,11 +176,7 @@ public struct MessageCenterListView: View {
                     .appearanceTint()
                     .opacity(1.0 - self.listOpacity)
             } else if self.messageIDs.isEmpty {
-                VStack {
-                    refreshButton()
-                    Text("ua_empty_message_list".messageCenterLocalizedString)
-                        .opacity(1.0 - self.listOpacity)
-                }
+                emptyMessageListMessage()
             }
         }
 
@@ -214,10 +212,10 @@ public struct MessageCenterListView: View {
                     Text(
                         "\("ua_delete_messages".messageCenterLocalizedString) (\(self.selection.count))"
                     )
-                    .foregroundColor(theme.deleteButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.deleteButtonTitleColorDark))
+                    .foregroundColor(colorScheme.resolveColor(light: theme.deleteButtonTitleColor, dark: theme.deleteButtonTitleColorDark))
                 } else {
                     Text("ua_delete_messages".messageCenterLocalizedString)
-                        .foregroundColor(theme.deleteButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.deleteButtonTitleColorDark))
+                        .foregroundColor(colorScheme.resolveColor(light: theme.deleteButtonTitleColor, dark: theme.deleteButtonTitleColorDark))
                 }
             }
         )
@@ -236,10 +234,10 @@ public struct MessageCenterListView: View {
                     Text(
                         "\("ua_mark_messages_read".messageCenterLocalizedString) (\(self.selection.count))"
                     )
-                    .foregroundColor(theme.markAsReadButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.markAsReadButtonTitleColorDark))
+                    .foregroundColor(colorScheme.resolveColor(light: theme.markAsReadButtonTitleColor, dark: theme.markAsReadButtonTitleColorDark))
                 } else {
                     Text("ua_mark_messages_read".messageCenterLocalizedString)
-                        .foregroundColor(theme.markAsReadButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.markAsReadButtonTitleColorDark))
+                        .foregroundColor(colorScheme.resolveColor(light: theme.markAsReadButtonTitleColor, dark: theme.markAsReadButtonTitleColorDark))
                 }
             }
         )
@@ -261,7 +259,7 @@ public struct MessageCenterListView: View {
             self.selection = Set(self.messageIDs)
         } label: {
             Text("ua_select_all_messages".messageCenterLocalizedString)
-                .foregroundColor(theme.selectAllButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.selectAllButtonTitleColorDark))
+                .foregroundColor(colorScheme.resolveColor(light: theme.selectAllButtonTitleColor, dark: theme.selectAllButtonTitleColorDark))
         }
         .accessibilityHint("ua_select_all_messages".messageCenterLocalizedString)
     }
@@ -271,7 +269,7 @@ public struct MessageCenterListView: View {
             self.selection = Set()
         } label: {
             Text("ua_select_none_messages".messageCenterLocalizedString)
-                .foregroundColor(theme.selectAllButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.selectAllButtonTitleColorDark))
+                .foregroundColor(colorScheme.resolveColor(light: theme.selectAllButtonTitleColor, dark: theme.selectAllButtonTitleColorDark))
         }
         .accessibilityHint("ua_select_none_messages".messageCenterLocalizedString)
     }
@@ -294,8 +292,8 @@ public struct MessageCenterListView: View {
         let isEditMode = self.editMode?.wrappedValue.isEditing ?? false
         let color =
             isEditMode
-        ? theme.cancelButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.cancelButtonTitleColorDark) :
-        theme.editButtonTitleColor?.adaptiveColor(for: colorScheme, darkVariation: theme.editButtonTitleColorDark)
+        ? colorScheme.resolveColor(light: theme.cancelButtonTitleColor, dark: theme.cancelButtonTitleColorDark) :
+        colorScheme.resolveColor(light: theme.editButtonTitleColor, dark: theme.editButtonTitleColorDark)
 
         return EditButton()
             .foregroundColor(color)
@@ -303,33 +301,65 @@ public struct MessageCenterListView: View {
     }
 
     @ViewBuilder
-    private func refreshButton() -> some View {
-        let refreshColor = theme.refreshTintColor?.adaptiveColor(for: colorScheme, darkVariation: theme.refreshTintColorDark)
+    private func toolbarRefreshButton() -> some View {
+        let refreshColor = colorScheme.resolveColor(light: theme.refreshTintColor, dark: theme.refreshTintColorDark)
 
-        if isRefreshing {
-            ProgressView()
-                .appearanceTint()
-        } else {
+        Button {
+            Task { @MainActor in
+                isRefreshing = true
+                await self.viewModel.refreshList()
+                isRefreshing = false
+            }
+        } label: {
+            if isRefreshing {
+                ProgressView()
+                    .appearanceTint()
+            } else {
+                Image(systemName: "arrow.clockwise")
+                    .foregroundColor(refreshColor)
+            }
+        }
+        .disabled(isRefreshing)
+    }
+
+    @ViewBuilder
+    private func emptyMessageListMessage() -> some View {
+        let refreshColor = colorScheme.resolveColor(light: theme.refreshTintColor, dark: theme.refreshTintColorDark)
+
+        VStack {
             Button {
-                Task {
+                Task { @MainActor in
                     isRefreshing = true
                     await self.viewModel.refreshList()
                     isRefreshing = false
                 }
             } label: {
-                Image(systemName: "arrow.clockwise")
-                    .foregroundColor(refreshColor)
+
+                ZStack {
+                    if isRefreshing {
+                        ProgressView()
+                            .appearanceTint()
+                    } else {
+                        Image(systemName: "arrow.clockwise")
+                            .foregroundColor(refreshColor)
+                    }
+                }
+                .frame(height: 44)
+                .background(Color.airshipTappableClear)
             }
             .disabled(isRefreshing)
-            .opacity(isRefreshing ? 0 : 1)
+
+            Text("ua_empty_message_list".messageCenterLocalizedString)
         }
+        .opacity(1.0 - self.listOpacity)
+
     }
 
     private func leadingToolbar() -> some ToolbarContent {
         ToolbarItemGroup(placement: .navigationBarTrailing) {
             editButton()
-            if #available(iOS 15, *) {} else {
-                refreshButton()
+            if #unavailable(iOS 15) {
+                toolbarRefreshButton()
             }
         }
     }